fix navigation and view chart in admin page

+++ vaccine_for_the_people/lib/app/data/models/sale_data.dart
@@ -1,5 +1,5 @@
 class SalesData {
   SalesData(this.year, this.sales);
   final String year;
-  final double sales;
+  final int sales;
 }
